no more " escaping, emit simple attr strings literally
Removed patch for "some\"thing" in attributes; it didn't work all
that well (it resulted in &#34; in the html, which was hideous
even if angular supported it), and single quotes work fine almost
all the time.

Simple attribute strings are now output literally instead of being
evaluated by the go template engine. In particular, this allows
semicolons in styles.